+++ packages/orq-rc/RELEASES.md
@@ -4049,3 +4049,13 @@ Based on:
 - [typescript v3.14.0-rc.4] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.4 - packages/orq-rc
+
+## 2025-10-02 15:14:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.633.2 (2.720.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.5] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.5 - packages/orq-rc